a few changes - birthday -> formatBirthdate and move to datetime utils - diagnosisDate -> formatDiagnosisDate and move to datetime utils - alphabetize format and datetime utils & tests - add a note in guidelines re: alphabetizing utils & tests
