+++ src/utils/format.js
@@ -28,7 +28,9 @@
  *
  * 2. Function naming scheme: the main verb here is `format`. Start all function names with that.
  *
- * 3. Try to be consistent in how params are used:
+ * 3. Function organizational scheme in this file and tests file: alphabetical plz
+ *
+ * 4. Try to be consistent in how params are used:
  * (e.g., always pass in `bgPrefs`) rather than some (subset) of bgUnits and/or bgBounds
  * and try to copy & paste JSDoc @param descriptions for common params.
  *
@@ -36,26 +38,8 @@
 
 import _ from 'lodash';
 import { format } from 'd3-format';
-// using d3-time-format because time is time of data access in
-// user’s browser time, not PwD’s configured timezone
-import { utcFormat } from 'd3-time-format';
 import { convertToMmolL } from './bloodglucose';
 import { BG_HIGH, BG_LOW, MMOLL_UNITS } from './constants';
-
-/**
- * formatDecimalNumber
- * @param {Number} val - numeric value to format
- * @param {Number} [places] - optional number of decimal places to display;
- *                            if not provided, will display as integer (0 decimal places)
- *
- * @return {String} numeric value rounded to the desired number of decimal places
- */
-export function formatDecimalNumber(val, places) {
-  if (places === null || places === undefined) {
-    return format('d')(val);
-  }
-  return format(`.${places}f`)(val);
-}
 
 /**
  * formatBgValue
@@ -93,27 +77,16 @@ export function formatBgValue(val, bgPrefs, outOfRangeThresholds) {
 }
 
 /**
- * birthday
- * @param  {Object} patient   the patient object that contains the profile
- * @return {String}           MMM D, YYYY formated birthday or empty if none
+ * formatDecimalNumber
+ * @param {Number} val - numeric value to format
+ * @param {Number} [places] - optional number of decimal places to display;
+ *                            if not provided, will display as integer (0 decimal places)
+ *
+ * @return {String} numeric value rounded to the desired number of decimal places
  */
-export function birthday(patient) {
-  const bday = _.get(patient, ['profile', 'patient', 'birthday'], '');
-  if (bday) {
-    return utcFormat('%b %-d, %Y')(Date.parse(bday));
+export function formatDecimalNumber(val, places) {
+  if (places === null || places === undefined) {
+    return format('d')(val);
   }
-  return '';
-}
-
-/**
- * diagnosisDate
- * @param  {Object} patient   the patient object that contains the profile
- * @return {String}           MMM D, YYYY formated diagnosisDate or empty if none
- */
-export function diagnosisDate(patient) {
-  const diagnosis = _.get(patient, ['profile', 'patient', 'diagnosisDate'], '');
-  if (diagnosis) {
-    return utcFormat('%b %-d, %Y')(Date.parse(diagnosis));
-  }
-  return '';
+  return format(`.${places}f`)(val);
 }
